+++ test/aoj/grl/bipartite_matching.2.test.cpp
@@ -1,11 +1,11 @@
 // competitive-verifier: PROBLEM https://onlinejudge.u-aizu.ac.jp/problems/GRL_7_A
 #include <iostream>
-#include "flow/hopcroft_karp.hpp"
+#include "flow/bipartite_matching.hpp"
 
 int main(void) {
     int l, r, m;
     std::cin >> l >> r >> m;
-    HopcroftKarp g(l, r);
+    BipartiteMatching g(l, r);
     while (m--) {
         int u, v;
         std::cin >> u >> v;
